Implement merge algorithm.
Rather than adding new sequence to cluster of first close canonical
sequence encountered, merge all clusters which contain a close
canonical sequence and add to merged cluster.

I thought this would reduce the number of singleton clusters found at
the end of the run but that doesn't seeme to be true.